You can move with right click and use SkillOne with the middle mouse button.
Originally, the mouse left click was going to be the "active" skill button that I was going to implement but there are some issues with it. If you hit the inventory button then click on resume, Bernard will instantly attack due to the left click on resume. Same issue with the level up dialog boxes. This is very annoying when you got enemies around because you are wasting turns.